requirements.txt
altair==5.1.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client==8.3.1
jupyter-core==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mysql-connector-python==8.0.31
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
streamlit==1.27.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.0
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.17.0
